Move database file
Move database file into .utility to avoid having tons of useless directories.
